Moved yark project files into its own project package, `yark`
Dependency relations on other recent packages are to be checked, and also some code
